Background transparency support
Reason for change:
Setting transparentBackground property.
It's not clrear with this property as by XRE protocol it should be
a subproperty of "options" property but I was told that it's going to
be a separate property and haven't found its usage in server code so
done as a separate.
Test Procedure:
haven't found a test in test suite or an
app setting that property, will probably need to add it to test suite
Risks: low
Signed-off-by: Yaroslav Zlotchenko <yzlotchenko@productengine.com>

Change-Id: Ib722075ee8b7d288b62e5203585a3b39e561a055