RDK-20517, RDK-20302: Store indexed db to realtime partition of xfs (XG1v4)
Reason for change: In HDD enabled devices with XFS, by default files will
be stored in the data partition, which is smaller in size.
We want to be able to store it to the bigger partition i.e. the realtime
partition of XFS.
Creating a systemd service which does the following:
1) Determine the XFS mount path.
2) Remove the database folder in XFS partition, if present already.
3) Recreate the database folder in XFS partition.
4) Set the real-time flag using the command
   xfs_io -c 'chattr +t' "$XFS_INDEXEDDB_PATH"

Test Procedure: Ran indexed db test and compared the results of
du -d 1 -xh /opt/data/
df -kh | grep sda2

Risks: Low

Change-Id: I55bba2ffe2240335b819162002c9a193448a55aa
Signed-off-by: vani_hegde@comcast.com